requirements.txt
# The requirements.txt file contains all the dependencies of the project

flask
Flask-SQLAlchemy
sqlalchemy
python-dotenv
Flask-Migrate
psycopg2-binary
flask-bcrypt
flask-cors
Flask-JWT-Extended

# Short explanation of the dependencies:

# flask is the Python web framework we are gonna use.

# psycopg2-binary is the driver to make the connection with the Postgres database.

# Flask-SQLAlchemy is the ORM to make the queries to the database.

# Flask-Migrate a SQLAlchemy database migrations for Flask applications using Alembic.

# flask-bcrypt to encrypt and compare the password during signup and login process

# Flask-JWT-Extended for handling JWT token during signup and login process
